chore: update all configs to mainnet
