+++ contracts-abi/config/testnet.go
@@ -1,21 +1,5 @@
 package config
 
-type Contracts struct {
-	BidderRegistry    string
-	ProviderRegistry  string
-	PreconfManager    string
-	Oracle            string
-	BlockTracker      string
-	SettlementGateway string
-}
-
-type L1Contracts struct {
-	ValidatorOptInRouter string
-	VanillaRegistry      string
-	MevCommitAVS         string
-	L1Gateway            string
-}
-
 var TestnetContracts = Contracts{
 	// If these addresses change for a testnet deployment,
 	// please also update snippets/testnet-addresses.mdx
