+++ infrastructure/nomad/playbooks/variables/environments.yml
@@ -17,5 +17,5 @@ environments:
     profile: "{{ profile }}"
     version: "{{ version }}"
     secrets: fetch
-    genesis_timestamp: "0x676d7ae2"
+    genesis_timestamp: "0x6772e470"
     domain: mev-commit.xyz
